requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.3.0
blinker==1.9.0
blis==1.3.0
boto3==1.40.18
botocore==1.40.18
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.0
cryptography==46.0.4
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
dateparser==1.2.2
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
faiss-cpu==1.11.0.post1
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
filelock==3.20.0
filetype==1.2.0
Flask==3.1.0
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.48.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.62.0
google-generativeai==0.8.6
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-oauth==0.15.1
httpx-sse==0.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langchain==0.3.27
langchain-community==0.3.27
langchain-core==1.2.9
langchain-google-genai==4.2.0
langchain-text-splitters==0.3.9
langcodes==3.5.0
langgraph==0.4.9
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.4.1
language_data==1.3.0
llvmlite==0.45.1
lxml==6.0.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.32
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.5.1
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numba==0.62.1
numpy==2.1.1
oauthlib==3.3.1
openai==1.97.1
openai-whisper==20250625
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
openpyxl==3.1.5
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
pillow==10.4.0
platformdirs==4.5.0
plotly==6.2.0
polars==1.37.1
polars-runtime-32==1.37.1
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.6
psutil==7.2.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.3
pyngrok==7.5.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.34.2
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
streamlit==1.46.1
streamlit-calendar==1.4.0
streamlit-oauth==0.1.14
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.4.7
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.0
vsdx==0.5.19
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
